+++ src/Security/Security/ChangeLog.md
@@ -19,6 +19,10 @@
 -->
 
 ## Upcoming Release
+* Added Alerts Suppression Rules to cmdlets:
+    `Get-AlertsSuppressionRule`
+    `Remove-AlertsSuppressionRule`
+    `Set-AlertsSuppressionRule`
 
 ## Version 1.1.1
 * Updated Security .NET SDK package reference to version 3.0.0